+++ src/table/palettes.ts
@@ -1,0 +1,42 @@
+export const HEAVY:string = [
+  '┏━┳┓',
+  '┃ ┃┃',
+  '┣━╋┫',
+  '┃ ┃┃',
+  '┗━┻┛'
+].join('')
+export const DEFAULT:string = [
+  '┏━┳┓',
+  '┃ ┃┃',
+  '┡━╇┩',
+  '│ ││',
+  '└─┴┘'
+].join('')
+export const THIN:string = [
+  '┌─┬┐',
+  '│ ││',
+  '├─┼┤',
+  '│ ││',
+  '└─┴┘'
+].join('')
+export const ROUNDED:string = [
+  '╭─┬╮',
+  '│ ││',
+  '├─┼┤',
+  '│ ││',
+  '╰─┴╯'
+].join('')
+export const SEMI_ROUNDED:string = [
+  '┏━┳┓',
+  '┃ ┃┃',
+  '┡━╇┩',
+  '│ ││',
+  '╰─┴╯'
+].join('')
+export const ASCII:string = [
+  '+-++',
+  '| ||',
+  '+-++',
+  '| ||',
+  '+-++'
+].join('')