+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Tables (fmt.table)
+
+## [0.2.2] - TBD
+- Separate methods from index
 
 ## [0.2.1] - 2019-10-09
 ### Added